Budget_System--Processing/cart-scraper/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
playwright==1.41.0
playwright-stealth==2.0.1
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.26.0
